third_party/blink/web_tests/virtual/deprecated-css-appearance-values-enabled/css-parser/deprecated-css-appearance-values-expected.txt

CONSOLE WARNING: The keyword 'inner-spin-button' specified to an 'appearance' property is not standardized. It will be removed in the future.
CONSOLE WARNING: The keyword 'push-button' specified to an 'appearance' property is not standardized. It will be removed in the future.
CONSOLE WARNING: The keyword 'square-button' specified to an 'appearance' property is not standardized. It will be removed in the future.
CONSOLE WARNING: The keyword 'slider-horizontal' specified to an 'appearance' property is not standardized. It will be removed in the future.
CONSOLE WARNING: The keyword 'searchfield-cancel-button' specified to an 'appearance' property is not standardized. It will be removed in the future.
CONSOLE WARNING: The keyword 'media-slider' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'media-sliderthumb' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'media-volume-slider' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'media-volume-sliderthumb' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'sliderthumb-horizontal' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'sliderthumb-vertical' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
